Create IE_Bookmarks-To_Chrome
This script dumps the Favorites folder for the user that runs the script, onto the bookmarks bar in chrome for that same user.
Any subfolder structure will be retained to one recursive step. Everything lower than that is merged up.
Make sure to set the name for the imported folder
No admin needed.