Create Email
This script was designed with the purpose of opening edge to a specific website. Taking a screenshot of that.
Pasting that screenshot into Word and saving it. Then using that word document as the attachment of an email.
the attachment is then moved to a sub-folder with a new name for archival purposes.

If outlook is not in the path variable (It should be) it will be searched for. If outlook is NOT in path variable AND multiple versions of outlook exist on machine. Email will not send. Instead you will get a message prompt with the paths to all the found outlooks and where the file is now located

There is a ctrl+enter sent to outlook, near end of script. Depending on your outlook settings you may need to modify this line. If you have it set to prompt for confirmation, a shift tab then another enter can be added to fully automate the sending of the email. I prefer to give it a manual look over and hit send myself.