Create Start-Menu_Pinning
This function allows automation for pinning items to start menu.
After the function there are a few parts there.
The first part is to make a list of currently pinned items, this is only needed if you wish to remove en mass the current start menu.
That array is saved and in the last part added back in.

The middle parts are two variables then when populated as comma separated arrays will then be called later on and add (or remove) any provided values. You MUST match the name as it appears on the start menu. For example, Microsoft Office Word. Depending on the version it may appear as Microsoft Word 20XX or for O365 it just appears as Word. If you put items that don't exist they get ignored with no errors that I have noticed.

When adding items they are put in a new group at the very top, this is the purpose of the bulk commented out sections. They will prevent the Windows auto grouping. It might make sense to run a test with a few apps with and without the commented out sections to see for yourself. The $pinnedapp array takes a little time to parse and populate so it does increase run time noticeably. From approximately 4 or 5 seconds up to 40 - 60 seconds