Update SoftwareQuery
I plan to make this function based. Where you call the script with a parameter either Path, Search (if Search than Path must be provided also), Name, or IN. Then it will spit out the right method based on what you provide.
Maybe I'll try to also pass in format table as a format parameter or something like that. If not I'll default it to the last line of the Method 3a section.