Update Outlook (advanced mail)
This script was made for reasons.
That included the need of including a screenshot pasted into a word document and using that in an email every week, the included date was mainly for organizational and archiving purposes. So I placed something similar in the email message body because I can

There are many ways to grab a file as an attachment I included another example of $(gci C:\Path\to\folder\containing\file -File | sort-object -Property LastWriteTime -Descending)[0] this will take the most recent FILE in a folder and use that instead. You could also go another step and similarly get rid of the -File and zip up a folder and use that created zip as the attachment.
IF NO ATTACHMENT will be used. You MUST change /a $attachment" to "/c ipm.note

There's a bit going on in this script. If your needs are different (slightly or a lot) send me a comment or message or something if you need help modifying this for your purposes.