Create Local_Admin-Password_Changer
Script to generate a complex random password. AND change the password of a LOCAL admin account.
BEFORE USING THIS, you MUST add something to send this password somewhere. If not Local admin password WILL be changed, and you will NOT know what that password is.
Change the LocalUserName variable or use it as a parameter
This script is based off of: https://cyber-defense.sans.org/blog/2013/08/01/reset-local-administrator-password-automatically-with-a-different-password-across-the-enterprise
The methodology mentioned there is practical and secure. Plus not too difficult to setup.