Create Get-ADUserPasswordInfo
Get a bunch of info on a user and their account\password. I sometimes even do this just to confirm the username is the one I am thinking of. The SID is helpful for HKEY_Users stuff (not HKCU)

Also alias to Get-Who since typing that out even with alt tab got annoying after a while