Create  Get-SqlQuery
Run SQL Query Wrapped in PowerShell, assumes you have access to that database, and is using windows authentication. Nothing baked into the script for username\password authentication to the database itself. That's probably something I can proabbly easily add later if needed as optional variables that are used if specified and defaults to windows if ignored.